move BitrateManager and TimeUtils to common module
