+++ common/src/main/java/com/pedro/common/Extensions.kt
@@ -1,0 +1,14 @@
+package com.pedro.common
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+/**
+ * Created by pedro on 3/11/23.
+ */
+
+suspend fun onMainThread(code: () -> Unit) {
+  withContext(Dispatchers.Main) {
+    code()
+  }
+}
