+++ common/src/test/java/com/pedro/common/util/Utils.kt
@@ -14,16 +14,25 @@
  * limitations under the License.
  */
 
-package com.pedro.srt.utils
+package com.pedro.common.util
+
+import org.mockito.MockedStatic
 
 /**
- * Created by pedro on 30/8/23.
+ * Created by pedro on 1/9/23.
  */
-object TimeUtils {
+object Utils {
 
-  @JvmStatic
-  fun getCurrentTimeMicro(): Long = System.nanoTime() / 1000
-
-  @JvmStatic
-  fun getCurrentTimeMillis(): Long = System.currentTimeMillis()
+  suspend fun useStatics(statics: List<MockedStatic<out Any>>, callback: suspend () -> Unit) {
+    val list = statics.toMutableList()
+    if (list.isEmpty()) callback()
+    else if (list.size == 1) {
+      list[0].use {
+        callback()
+      }
+    } else {
+      val value = list.removeAt(0)
+      value.use { useStatics(list, callback) }
+    }
+  }
 }
