--- rtmp/src/main/java/com/pedro/rtmp/utils/TimeUtils.kt
@@ -1,29 +1,0 @@
-/*
- * Copyright (C) 2023 pedroSG94.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.pedro.rtmp.utils
-
-/**
- * Created by pedro on 30/8/23.
- */
-object TimeUtils {
-
-  @JvmStatic
-  fun getCurrentTimeMillis(): Long = System.currentTimeMillis()
-
-  @JvmStatic
-  fun getCurrentTimeNano(): Long = System.nanoTime()
-}
