+++ srt/src/main/java/com/pedro/srt/srt/SrtSender.kt
@@ -18,7 +18,9 @@ package com.pedro.srt.srt
 
 import android.media.MediaCodec
 import android.util.Log
+import com.pedro.common.BitrateManager
 import com.pedro.common.ConnectChecker
+import com.pedro.common.onMainThread
 import com.pedro.srt.mpeg2ts.Codec
 import com.pedro.srt.mpeg2ts.MpegTsPacket
 import com.pedro.srt.mpeg2ts.MpegTsPacketizer
@@ -31,9 +33,7 @@ import com.pedro.srt.mpeg2ts.psi.TableToSend
 import com.pedro.srt.mpeg2ts.service.Mpeg2TsService
 import com.pedro.srt.srt.packets.SrtPacket
 import com.pedro.srt.srt.packets.data.PacketPosition
-import com.pedro.srt.utils.BitrateManager
 import com.pedro.srt.utils.SrtSocket
-import com.pedro.srt.utils.onMainThread
 import com.pedro.srt.utils.trySend
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
