+++ srt/src/test/java/com/pedro/srt/mpeg2ts/PesTest.kt
@@ -16,8 +16,8 @@
 
 package com.pedro.srt.mpeg2ts
 
+import com.pedro.common.TimeUtils
 import com.pedro.srt.Utils
-import com.pedro.srt.utils.TimeUtils
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertArrayEquals
 import org.junit.Before
